修改了部分bug，编码器接反了，修改了OLED的界面

+++ App/inductance.c
@@ -88,7 +88,7 @@ void Get_AD_Value()
 	for (i = 0; i < 2; i++)											//计算差比和后的电感值
 	{
 		temp = Road_Data[i].AD_Value;
-		Road_Data[i].Normalized_Value = (temp / Direction.Normalization_Value) * 200;
+		Road_Data[i].Normalized_Value = (int)((temp / (double)Direction.Normalization_Value) * 200.0);
 	}
 }
